Added notifications actor that should encapsulate all other possible notifications actors. #232
